requirements.txt
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.4
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.37
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0

Pillow~=9.2.0
python-dotenv~=1.0.0